+++ streetSoccerManager-Backend/streetSoccerManager/src/main/java/soccer/game/streetSoccerManager/model/PlayerStats.java
@@ -7,24 +7,12 @@ import lombok.Setter;
 @Setter
 public class PlayerStats {
 
-    private int id;
+    private Long id;
     private int overallRating;
 
-    public PlayerStats(int id, int overallRating) {
+    public PlayerStats(Long id, int overallRating) {
         this.id = id;
         this.overallRating = overallRating;
     }
-
-    //    private int physical;
-//    private int defence;
-//    private int shooting;
-//    private int goalkeeping;
-//
-//    private int pace;
-//    private int passing;
-//    private int dribbling;
-//    private int defending;
-
-
 
 }
