+++ streetSoccerManager-Backend/streetSoccerManager/src/main/java/soccer/game/streetSoccerManager/repository/repositoryInterfaces/ITeamRepository.java
@@ -7,8 +7,8 @@ import java.util.List;
 
 public interface ITeamRepository {
     List<Team> getAll();
-    Team get(int id);
-    Boolean delete(int id);
+    Team get(Long id);
+    Boolean delete(Long id);
     Boolean add(Team team);
     Boolean update(Team team);
